+++ SourceryRuntime/Sources/JSExport.generated.swift
@@ -1,4 +1,4 @@
-// Generated using Sourcery 0.7.2 — https://github.com/krzysztofzablocki/Sourcery
+// Generated using Sourcery 0.8.0 — https://github.com/krzysztofzablocki/Sourcery
 // DO NOT EDIT
 
 import JavaScriptCore
@@ -142,7 +142,6 @@ extension Enum: EnumAutoJSExport {}
 
 extension EnumCase: EnumCaseAutoJSExport {}
 
-
 @objc protocol MethodAutoJSExport: JSExport {
     var name: String { get }
     var selectorName: String { get }
@@ -225,7 +224,6 @@ extension MethodParameter: MethodParameterAutoJSExport {}
 }
 
 extension Protocol: ProtocolAutoJSExport {}
-
 
 @objc protocol StructAutoJSExport: JSExport {
     var kind: String { get }
@@ -346,8 +344,6 @@ extension Type: TypeAutoJSExport {}
 
 extension TypeName: TypeNameAutoJSExport {}
 
-
-
 @objc protocol TypesCollectionAutoJSExport: JSExport {
 }
 
@@ -375,5 +371,3 @@ extension TypesCollection: TypesCollectionAutoJSExport {}
 }
 
 extension Variable: VariableAutoJSExport {}
-
-