+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 ## Master
 
+### New Features
+
+- Added support for file paths in `config` parameter
+- Improved config file validation and error reporting
+
 
 ## 0.8.0